requirements.txt
ccxt>=4.4.0
pandas>=2.0.0
pandas_ta>=0.3.14b0
PyYAML>=6.0.0
